handle next page

+++ file.py
@@ -1,8 +1,13 @@
 import os
+from pathlib import Path
 
 def create_file(file_name, file_content):
   path = os.path.abspath('.') + "/data/" + file_name
   print(path)
-  file = open(path, "w")
+  if Path(path).is_file():
+    file = open(path, "a")
+  else:
+    file = open(path, "w")
   file.write(file_content)
+  file.write('\n\r')
   file.close()
